Update from comments on the PR
